Setup root to be able to SSH in.
This is enough that there is a zone ready to be used as a new system!

We can now use the hostname.home.philipcristiano.com to get to a zone,
root can login to configure it, and it will be usable! 

Still need to move the `test_zone` module into a reusable one that can
just take a zone name.